uint16 as ms - bad bug and some acu checks need to be looked over

+++ Application/Src/stateMachine.c
@@ -22,7 +22,7 @@ volatile StatusLump globalStatus = {
 
 volatile uint8_t numberOfBadMessages = 0;
 int32_t dischargeStartMillis = BAD_TIME_Negative1;
-uint16_t lastECUStatusMsgTick = 0;
+uint32_t lastECUStatusMsgTick = 0;
 
 static const uint16_t howOftenToSpamECUStatusMsgs = 35;
